Separate logger into new autoload
