+++ Framework/Autoloads/Global/Global.cs
@@ -1,7 +1,6 @@
 using Godot;
 using System.Threading.Tasks;
 using System;
-using __TEMPLATE__.UI;
 
 namespace __TEMPLATE__;
 
@@ -9,23 +8,10 @@ public partial class Global : Node
 {
     public event Func<Task> PreQuit;
 
-    public static Global Instance { get; private set; }
-
-    public Logger Logger { get; private set; } = new();
-
-    public override void _Ready()
-    {
-        Instance = this;
-        
-        Logger.MessageLogged += GetNode<UIConsole>(Autoloads.Console).AddMessage;
-
-        //ModLoaderUI.LoadMods(this);
-    }
-
-    public override void _PhysicsProcess(double delta)
-    {
-        Logger.Update();
-    }
+    //public override void _Ready()
+    //{
+    //    ModLoaderUI.LoadMods(this);
+    //}
 
     public override async void _Notification(int what)
     {
