+++ Framework/Game.cs
@@ -13,21 +13,21 @@ public static class Game
 
     public static void Log(object message, BBColor color = BBColor.Gray)
     {
-        Global.Instance.Logger.Log(message, color);
+        LoggerManager.Instance.Logger.Log(message, color);
     }
 
     public static void Log(params object[] objects)
     {
-        Global.Instance.Logger.Log(objects);
+        LoggerManager.Instance.Logger.Log(objects);
     }
 
     public static void LogWarning(object message, BBColor color = BBColor.Orange)
     {
-        Global.Instance.Logger.LogWarning(message, color);
+        LoggerManager.Instance.Logger.LogWarning(message, color);
     }
 
     public static void LogErr(Exception e, string hint = null)
     {
-        Global.Instance.Logger.LogErr(e, hint);
+        LoggerManager.Instance.Logger.LogErr(e, hint);
     }
 }
