+++ project.godot
@@ -22,6 +22,7 @@ ServiceProvider="*res://Framework/Autoloads/Services/Services.cs"
 OptionsManager="*res://Framework/Autoloads/OptionsManager.cs"
 Console="*res://Framework/Autoloads/Console.tscn"
 Global="*res://Framework/Autoloads/Global/Global.tscn"
+LoggerManager="*res://Framework/Autoloads/LoggerManager.cs"
 AudioManager="*res://Framework/Autoloads/AudioManager/AudioManager.tscn"
 SceneManager="*res://Framework/Autoloads/SceneManager/SceneManager.tscn"
 VisualizeAutoload="*res://Framework/GodotUtils/Visualize/Scripts/VisualizeAutoload.cs"
